requirements-local.txt
# requirements-local.txt
flask
flask-cors
pymongo
python-dotenv